Lessons: Note2Position, Position2Note. Correct enabling/disabling of views.
The two lessons mention above are implemented as Proof-of-Concept. The lessons enable views,
which are required by them. So the Note2Position enable only NotesView and FretView. The user
in this lesson is shown a note and must input it in the FretView.

The NoteView select can now be done with gestures. Thus the problem is observed in the current
implementation: the long touch and swiping is not recognized as gesture, the short touch and swipe
is required.